requirements.txt
python-dotenv==1.0.0
pymongo==4.3.3
requests==2.25.1
certifi==2022.12.7
boto3==1.26.159